requirements.txt
# pytest
pytest==7.4.2
pytest-xdist==3.3.1
pytest-custom_exit_code==0.3.0
pytest-rerunfailures==11.1.2
# framework extras
selenium==4.13.0
omegaconf==2.3.0 
requests==2.31.0
allure-pytest==2.13.2
strenum==0.4.10
# additional packets
openpyxl==3.1.2
six

